% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svregrp.R
\name{svregrp_Gibbs_area}
\alias{svregrp_Gibbs_area}
\title{Gibbs sampling of the grouped household effect model with area random effect}
\usage{
svregrp_Gibbs_area(
  Y_star,
  x_covs,
  z_covs,
  i_ind,
  area_ind,
  sh_ind,
  hit_ind,
  max_steps,
  cor_step_size,
  corr_vs_diag = FALSE,
  calcu_DIC = TRUE,
  burn_in = 1,
  verbose = TRUE,
  init_mean_coeffs = NULL,
  init_corr_coeffs = NULL,
  init_sigma2_e = NULL,
  init_sigma2_u = NULL,
  init_sigma2_v = NULL,
  init_sigma2_w = NULL
)
}
\arguments{
\item{Y_star}{A NN * 1 matrix containing continuous responses.}

\item{x_covs}{A NN * p matrix containing covariates for the mean model.}

\item{z_covs}{A NN * q matrix containing covariates for the correlation model.}

\item{i_ind}{A vector containing indices for individuals.}

\item{area_ind}{A vector containing indices for areas.}

\item{sh_ind}{A vector containing indices for super-households.}

\item{hit_ind}{A vector containing indices for unique households.}

\item{max_steps}{Length of MCMC chains to be drawn.}

\item{cor_step_size}{A vector length q specifying step sizes for q coefficients
of correlation model.}

\item{corr_vs_diag}{Whether set the grouped household effects to be independent,
default value is FALSE.}

\item{calcu_DIC}{Whether to calculate DIC values, default value is TRUE.}

\item{burn_in}{Burn in size for MCMC, default value is 1.}
}
\value{
A data frame with columns of MCMC chains of sampled parameters and latent variables.
}
\description{
This function takes response, multilevel indices as input and
returns MCMC sample chains of model parameters and latent variables.
}
