% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svregrp.R
\name{svregrp_Gibbs_area_nohe}
\alias{svregrp_Gibbs_area_nohe}
\title{Gibbs sampling of the grouped household effect model with area random effect (remove the household effect)}
\usage{
svregrp_Gibbs_area_nohe(
  Y_star,
  x_covs,
  i_ind,
  area_ind,
  max_steps,
  calcu_DIC = TRUE,
  burn_in = 1,
  verbose = TRUE,
  init_mean_coeffs = NULL,
  init_sigma2_e = NULL,
  init_sigma2_u = NULL,
  init_sigma2_w = NULL
)
}
\arguments{
\item{Y_star}{A NN * 1 matrix containing continuous responses.}

\item{x_covs}{A NN * p matrix containing covariates for the mean model.}

\item{i_ind}{A vector containing indices for individuals.}

\item{area_ind}{A vector containing indices for areas.}

\item{max_steps}{Length of MCMC chains to be drawn.}

\item{calcu_DIC}{Whether to calculate DIC values, default value is TRUE.}

\item{burn_in}{Burn in size for MCMC, default value is 1.}
}
\value{
A data frame with columns of MCMC chains of sampled parameters and latent variables.
}
\description{
This function takes response, multilevel indices as input and
returns MCMC sample chains of model parameters and latent variables.
}
