% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svregrp.R
\name{svregrp_Gibbs_mchains}
\alias{svregrp_Gibbs_mchains}
\title{Multiple chain Gibbs sampling for grouped household effect model}
\usage{
svregrp_Gibbs_mchains(
  Y_star,
  x_covs,
  z_covs,
  i_ind,
  sh_ind,
  hit_ind,
  max_steps,
  cor_step_size,
  corr_vs_diag = FALSE,
  num_chains = 1,
  output_progress_file = "output_chain_1.txt",
  init_mean_coeffs = NULL,
  init_corr_coeffs = NULL,
  init_sigma2_e = NULL,
  init_sigma2_u = NULL,
  init_sigma2_v = NULL
)
}
\arguments{
\item{Y_star}{A NN * 1 matrix containing continuous responses.}

\item{x_covs}{A NN * p matrix containing covariates for the mean model.}

\item{z_covs}{A matrix containing covariates for the correlation model.}

\item{i_ind}{A vector containing indices for individuals.}

\item{sh_ind}{A vector containing indices for super-households.}

\item{hit_ind}{A vector containing indices for households.}

\item{max_steps}{Length of MCMC chains to be drawn.}

\item{cor_step_size}{Step sizes for correlation parameter updates.}

\item{corr_vs_diag}{Whether to use diagonal correlation structure.}

\item{num_chains}{Number of parallel chains to run.}

\item{output_progress_file}{File name for progress output.}

\item{init_mean_coeffs}{Initial values for mean coefficients.}

\item{init_corr_coeffs}{Initial values for correlation coefficients.}

\item{init_sigma2_e}{Initial value for error variance.}

\item{init_sigma2_u}{Initial value for individual random effect variance.}

\item{init_sigma2_v}{Initial value for household random effect variance.}
}
\value{
A list containing multiple MCMC chains as mcmc.list object for convergence diagnostics.
}
\description{
This function runs multiple MCMC chains for convergence diagnostics and
provides more robust parameter estimation through multiple independent chains.
}
